rename generate to regional
